feat: Implement HTX exchange services for spot, super-margin, and swap accounts
- Added services for fetching spot account information, super margin account information, and swap account information.
- Implemented order listing and submission for swap and super-margin accounts.
- Created market data services for interest rates and quotes.
- Provided product listing services for swap and spot products.
- Established UID caching for account identification.
- Removed deprecated transfer functionalities.
