+++ apps/vendor-huobi/package.json
@@ -22,6 +22,7 @@
     "@yuants/data-series": "workspace:*",
     "@yuants/data-interest-rate": "workspace:*",
     "@yuants/data-quote": "workspace:*",
+    "@yuants/exchange": "workspace:*",
     "rxjs": "~7.5.6",
     "crypto-js": "^4.2.0"
   },
