+++ common/changes/@yuants/vendor-huobi/2025-11-26-07-24.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@yuants/vendor-huobi",
+      "comment": "renew HTX",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@yuants/vendor-huobi"
+}
